requirements.txt
streamlit
torch
torchvision
pillow
numpy
pandas
google-generativeai
requests
firebase-admin>=6.2.0
google-cloud-firestore>=2.13.0
google-auth-oauthlib
python-dateutil>=2.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0
streamlit-geolocation